Move App component to the components folder
