+++ src/components/App/App.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { render } from 'testUtils';
-import App from 'App';
+import App from '.';
 
 describe('App', () => {
   it('should render without crashing', () => {
