+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM, { render } from 'react-dom';
 import { Provider } from 'react-redux';
 import { PersistGate } from 'redux-persist/integration/react';
 
-import App from 'App';
+import App from 'components/App';
 import { persistor, store } from 'store';
 import * as serviceWorker from 'serviceWorker';
